0827 Security

+++ src/main/java/com/example/k5_iot_springboot/provider/JwtProvider.java
@@ -53,7 +53,7 @@ public class JwtProvider {
     /** 서명용 비밀키, 액세스 토큰 만료시간(ms), 만료 직후 허용할 시계 오차(s) */
 //    환경 변수에 지정한 비밀키와 만료 시간 변수 선언
     private final SecretKey key;
-    private final int jwtExpirationMs;
+    private final long jwtExpirationMs;
     private final int clockSkewSeconds;
 
     /** 검증/파싱 파서: 파서를 생성자에서 1회 구성하여 재사용 - 성능/일관성 보장 (JJWT의 파서 객체) */
@@ -64,7 +64,7 @@ public class JwtProvider {
 //            @Value: application.properties나 application.yml과 같은 설정 파일의 값을 클래스 변수에 주입
 //                      >> 데이터 타입 자동 인식
             @Value("${jwt.secret}") String secret, // cf) Base64 인코딩된 비밀키 문자열이어야 함
-            @Value("${jwt.expiration}") int jwtExpirationMs,
+            @Value("${jwt.expiration}") long jwtExpirationMs,
             @Value("${jwt.clock-skew-seconds:0}") int clockSkewSeconds // 기본 0 - 옵션
     ) {
 //        생성자: JwtProvider 객체 생성 시 비밀키와 만료시간 초기화
